- Update ContactDAO.kt so it works properly with annotations - Fix ContactDatabase.kt - Fix ContactRepository.kt - Add view parameter method in ContactViewModel.kt - Add card_item.xml - Finish activity_main.xml - Create MyRecyclerViewAdapter.kt - Create ViewModelFactory.kt - Initialize ROOM DB in MainActivity.kt, ViewModel and recycler view

+++ app/src/main/java/com/example/contactsmanagerkotlinapp/repository/ContactRepository.kt
@@ -11,7 +11,7 @@ class ContactRepository(private val contactDAO: ContactDAO) {
         return contactDAO.insertContact(contact)
     }
 
-    suspend fun delete(contact: Contact): Long {
+    suspend fun delete(contact: Contact): Int {
         return contactDAO.deleteContact(contact)
     }
 
@@ -19,7 +19,7 @@ class ContactRepository(private val contactDAO: ContactDAO) {
         return contactDAO.deleteAllContacts()
     }
 
-    suspend fun update(contact: Contact): Long {
+    suspend fun update(contact: Contact): Int {
         return contactDAO.updateContact(contact)
     }
 }
